Move shift_free & shift_validate endpoints to ShiftController
