+++ app/Resources/views/admin/booking/_partial/bucket_modal.html.twig
@@ -40,20 +40,20 @@ It use the materialize modal class https://materializecss.com/modals.html
                     {% if is_granted('free',shift) %}
                         {% if use_card_reader_to_validate_shifts and shift.isPastOrCurrent %}
                             {% if shift.wasCarriedOut %}
-                                <form action="{{ path('invalidate_shift', { 'id' : shift.id }) }}" method="POST" id="invalidate_shift_{{ shift.id }}">
+                                <form action="{{ path('shift_invalidate', { 'id': shift.id }) }}" method="POST" id="shift_invalidate_{{ shift.id }}">
                                     <button type="submit" class="btn red">
                                         <i class="material-icons left">highlight_off</i>Invalider la participation
                                     </button>
                                 </form>
                             {% else %}
-                                <form action="{{ path('validate_shift', { 'id' : shift.id }) }}" method="POST" id="validate_shift_{{ shift.id }}">
+                                <form action="{{ path('shift_validate', { 'id': shift.id }) }}" method="POST" id="shift_validate_{{ shift.id }}">
                                     <button type="submit" class="btn green">
                                         <i class="material-icons left">check_circle</i>Valider la participation
                                     </button>
                                 </form>
                             {% endif %}
                         {% endif %}
-                        <form action="{{ path('free_shift', { 'id' : shift.id }) }}" method="POST" id="free_shift_{{ shift.id }}">
+                        <form action="{{ path('shift_free', { 'id': shift.id }) }}" method="POST" id="shift_free_{{ shift.id }}">
                             {% if shift.isPast %}
                                 <button type="submit" class="btn red">
                                     <i class="material-icons left">delete</i>Supprimer
