+++ src/AppBundle/Controller/ShiftController.php
@@ -6,11 +6,15 @@ use DateTime;
 use AppBundle\Entity\Job;
 use AppBundle\Entity\Shift;
 use AppBundle\Event\ShiftBookedEvent;
+use AppBundle\Event\ShiftFreedEvent;
+use AppBundle\Event\ShiftValidatedEvent;
+use AppBundle\Event\ShiftInvalidatedEvent;
 use AppBundle\Form\ShiftType;
 use AppBundle\Security\MembershipVoter;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Session\Session;
 use Symfony\Component\Form\Extension\Core\Type\RadioType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
@@ -190,7 +194,126 @@ class ShiftController extends Controller
             $session->getFlashBag()->add("success", "Créneau réservé avec succès pour " . $shift->getShifter());
             return $this->redirectToRoute('booking_admin');
         }
+
         $session->getFlashBag()->add('error', "Une erreur est survenue...");
+        return $this->redirectToRoute('booking_admin');
+    }
+
+    /**
+     * free a shift.
+     *
+     * @Route("/{id}/free", name="shift_free")
+     * @Method("POST")
+     */
+    public function freeShiftAction(Request $request, Shift $shift)
+    {
+        $this->denyAccessUnlessGranted(ShiftVoter::FREE, $shift);
+
+        $session = new Session();
+
+        $membership = $shift->getShifter()->getMembership();
+
+        $em = $this->getDoctrine()->getManager();
+        $shift->free();
+        $shift->invalidateShiftParticipation();
+        $em->persist($shift);
+        $em->flush();
+
+        $dispatcher = $this->get('event_dispatcher');
+        $dispatcher->dispatch(ShiftFreedEvent::NAME, new ShiftFreedEvent($shift, $membership));
+
+        $session->getFlashBag()->add('success', "Le créneau a bien été libéré");
+
+        $referer = $request->headers->get('referer');
+        return new RedirectResponse($referer);
+
+    }
+
+    /**
+     * validate a shift.
+     *
+     * @Route("/{id}/validate", name="shift_validate")
+     * @Method("POST")
+     */
+    public function validateShiftAction(Request $request, Shift $shift)
+    {
+        $this->denyAccessUnlessGranted(ShiftVoter::VALIDATE, $shift);
+
+        $session = new Session();
+
+        if ($shift->getWasCarriedOut() == 0) {
+            $membership = $shift->getShifter()->getMembership();
+
+            $em = $this->getDoctrine()->getManager();
+            $shift->validateShiftParticipation();
+            $em->persist($shift);
+            $em->flush();
+
+            $dispatcher = $this->get('event_dispatcher');
+            $dispatcher->dispatch(ShiftValidatedEvent::NAME, new ShiftValidatedEvent($shift));
+
+            $session->getFlashBag()->add('success', "La participation au créneau a bien été validée");
+        } else {
+            $session->getFlashBag()->add('error', "La participation au créneau a déjà été validée");
+        }
+
+        $referer = $request->headers->get('referer');
+        return new RedirectResponse($referer);
+    }
+
+    /**
+     * invalidate a shift.
+     *
+     * @Route("/{id}/invalidate", name="shift_invalidate")
+     * @Method("POST")
+     */
+    public function invalidateShiftAction(Request $request, Shift $shift)
+    {
+        $this->denyAccessUnlessGranted(ShiftVoter::INVALIDATE, $shift);
+
+        $session = new Session();
+
+        if ($shift->getWasCarriedOut() == 1) {
+            $membership = $shift->getShifter()->getMembership();
+
+            $em = $this->getDoctrine()->getManager();
+            $shift->invalidateShiftParticipation();
+            $em->persist($shift);
+            $em->flush();
+
+            $dispatcher = $this->get('event_dispatcher');
+            $dispatcher->dispatch(ShiftInvalidatedEvent::NAME, new ShiftInvalidatedEvent($shift, $membership));
+
+            $session->getFlashBag()->add('success', "La participation au créneau a bien été invalidée");
+        } else {
+            $session->getFlashBag()->add('error', "La participation au créneau a déjà été invalidée");
+        }
+
+        $referer = $request->headers->get('referer');
+        return new RedirectResponse($referer);
+    }
+
+    /**
+     * remove a shift.
+     *
+     * @Route("/{id}", name="shift_delete")
+     * @Security("has_role('ROLE_SHIFT_MANAGER')")
+     * @Method("DELETE")
+     */
+    public function removeShiftAction(Request $request, Shift $shift)
+    {
+        $session = new Session();
+
+        $form = $this->createDeleteForm($shift);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($shift);
+            $em->flush();
+            $session->getFlashBag()->add('success', 'Le créneau a bien été supprimé !');
+        }
+
         return $this->redirectToRoute('booking_admin');
     }
 
@@ -225,30 +348,17 @@ class ShiftController extends Controller
     }
 
     /**
-     * remove a shift.
+     * Creates a form to delete a shift entity.
      *
-     * @Route("/{id}", name="shift_delete")
-     * @Security("has_role('ROLE_SHIFT_MANAGER')")
-     * @Method("DELETE")
+     * @param Shift $shift The shift entity
+     *
+     * @return \Symfony\Component\Form\Form The form
      */
-    public function removeShiftAction(Request $request, Shift $shift)
+    private function createDeleteForm(Shift $shift)
     {
-        $session = new Session();
-
-        $form = $this->createFormBuilder()
+        return $this->createFormBuilder()
             ->setAction($this->generateUrl('shift_delete', array('id' => $shift->getId())))
             ->setMethod('DELETE')
             ->getForm();
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($shift);
-            $em->flush();
-            $session->getFlashBag()->add('success', 'Le créneau a bien été supprimé !');
-        }
-
-        return $this->redirectToRoute('booking_admin');
     }
-
 }
